+++ people-and-teams/profilecard/examples/10-profilecard-inline-edit.tsx
@@ -1,6 +1,5 @@
 import React, { useState } from 'react';
 
-// Whaaaa
 import styled from 'styled-components';
 
 import Select, { ValueType } from '@atlaskit/select';